Adds "Inner Critic" EE on Wednesday (1.2.3)
Day 3 features a checkpoint assessments, it seems
like the pressure cooker has been on long enough
to have a check in about one's expectations for
oneself. Here's where voices about "repeat" and
"not good enough" fit in. I like it /after/ the
assessment because this message doesn't wind up
negatively priming the audience.